separated CharacterTable and Character components
