+++ components/character/character.module.css
@@ -1,0 +1,21 @@
+.characterRow {
+    margin-bottom: 10px;
+}
+
+.characterField, .profilePic {
+    padding-top: 10px;
+    padding-bottom: 10px;
+    vertical-align: middle;
+}
+
+.profilePic {
+    width:100%;
+}
+
+.profilePic > img {
+    height: 100px;
+    width: 100%;
+    display: inline-block;
+    border-radius: 50%;
+    object-fit: cover;
+}
